+++ src/components/common/SearchPanels/SearchPanelPrimary.jsx
@@ -1,9 +1,9 @@
 import { memo } from 'react';
-import { useSearchQuery } from '../../../hooks/useSearchQuery';
+import { useCitySearchQuery } from '../../../hooks/useCitySearchQuery';
 import SearchIcon from '../../../assets/img/searchIcon.svg';
 
 const SearchPanelPrimary = () => {
-  const { query, setQuery, updateQuery } = useSearchQuery();
+  const { query, setQuery, updateQuery } = useCitySearchQuery();
 
   const handleSubmit = event => {
     event.preventDefault();
